+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.0.0 --installs on--> 2.4 IDE-EE / 3.1 IDE-CE
 
+## Current
+
+- [FP-2261 - Added Open Callback option to Port Context Menu](https://movai.atlassian.net/browse/FP-2261)
+
 ## Miscellaneous
 
 - [FP-2027 - Editor shouldn't require method "renderRightMenu" implementation](https://movai.atlassian.net/browse/FP-2027)
@@ -14,5 +18,3 @@
 - [FP-2152 - Create new editor will close editor window](https://movai.atlassian.net/browse/FP-2152)
 - [FP-2128 - Add container configurations in Node editor](https://movai.atlassian.net/browse/FP-2147)
 - [FP-2186 - Main Menu tools are not showing up](https://movai.atlassian.net/browse/FP-2186)
-
-